Use proof queue asynchronously
